requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asgiref=3.3.4=pyhd3eb1b0_0
ca-certificates=2021.4.13=haa95532_1
certifi=2020.12.5=py39haa95532_0
django=3.2=pyhd3eb1b0_0
django-filter=2.4.0=pypi_0
djangorestframework=3.12.4=pypi_0
krb5=1.17.1=hc04afaa_0
libpq=12.2=h3235a2c_0
markdown=3.3.4=pypi_0
openssl=1.1.1k=h2bbff1b_0
pip=21.0.1=py39haa95532_0
psycopg2=2.8.6=py39hcd4344a_1
python=3.9.4=h6244533_0
pytz=2021.1=pyhd3eb1b0_0
setuptools=52.0.0=py39haa95532_0
sqlite=3.35.4=h2bbff1b_0
sqlparse=0.4.1=py_0
tk=8.6.10=he774522_0
tzdata=2020f=h52ac0ba_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py39h2bbff1b_0
zlib=1.2.11=h62dcd97_4
